gcp_prac/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
apache-beam[gcp]==2.37.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12; python_version >= '3'
cloudpickle==2.0.0; python_version >= '3.6'
crcmod==1.7
dill==0.3.1.1; python_version >= '2.6' and python_version != '3.0'
docopt==0.6.2
fastavro==1.4.10; python_version >= '3.7'
fasteners==0.17.3; python_version >= '3.6'
google-api-core[grpc]==1.31.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-apitools==0.5.31
google-auth==1.35.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-bigquery-storage==2.12.0
google-cloud-bigquery==2.34.1
google-cloud-bigtable==1.7.0
google-cloud-core==1.7.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-datastore==1.15.3
google-cloud-dlp==3.6.1
google-cloud-language==1.3.0
google-cloud-pubsub==2.10.0
google-cloud-pubsublite==1.4.0
google-cloud-recommendations-ai==0.2.0
google-cloud-spanner==1.19.1
google-cloud-storage==2.1.0
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-crc32c==1.3.0; python_version >= '3.6'
google-resumable-media==2.3.1; python_version >= '3.6'
googleapis-common-protos==1.55.0; python_version >= '3.6'
grpc-google-iam-v1==0.12.3
grpcio-gcp==0.2.2
grpcio-status==1.44.0
grpcio==1.44.0; python_version >= '3.6'
hdfs==2.6.0
httplib2==0.19.1
idna==3.3; python_version >= '3'
numpy==1.21.5
oauth2client==4.1.3
orjson==3.6.7; python_version >= '3.7'
overrides==6.1.0; python_version >= '3.6'
packaging==21.3; python_version >= '3.6'
pandas==1.4.1
proto-plus==1.20.3; python_version >= '3.6'
protobuf==3.19.4; python_version >= '3.5'
pyarrow==6.0.1; python_version >= '3.6'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydot==1.4.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pymongo==3.12.3
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.3
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rsa==4.8; python_version >= '3.6'
setuptools==60.9.3; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==4.1.1; python_version >= '3.6'
typing-utils==0.1.0; python_full_version >= '3.6.1'
urllib3==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
